+++ src/index.js
@@ -1,5 +1,6 @@
 import './editors/wrapper-editor.js';
 import './cards/wrapper-card.js';
+import './cards/title-card.js';
 import './cards/light-card.js';
 import './cards/shutter-card.js';
 import './cards/navigation-card.js';